Move initialization of http client into client hook
