--- src/main/frontend/src/lib/stores/auth_store.ts
@@ -1,13 +1,0 @@
-import { writable } from 'svelte/store';
-
-interface AuthStoreProps {
-  isAuthenticated: boolean;
-  roles: string[];
-  bearerToken: string;
-}
-
-export const AuthStore = writable<AuthStoreProps>({
-  isAuthenticated: false,
-  roles: [],
-  bearerToken: '',
-});
