+++ src/main/frontend/vite.config.ts
@@ -4,28 +4,33 @@ import { defineConfig } from 'vite';
 import Icons from 'unplugin-icons/vite';
 
 export default defineConfig({
-    plugins: [
-        Icons({
-            compiler: 'svelte',
-            autoInstall: true,
-        }),
-        sveltekit(),
-        tailwindcss()
-    ],
-    server: {
-        proxy: {
-            '/realms': {
-                target: 'http://127.0.0.1:9090',
-                changeOrigin: true
-            },
-            '/resources': {
-                target: 'http://127.0.0.1:9090',
-                changeOrigin: true
-            },
-            '/api': {
-                target: 'http://127.0.0.1:8080',
-                changeOrigin: true,
-            },
-        }
-    }
+	esbuild: {
+		supported: {
+			'top-level-await': true
+		}
+	},
+	plugins: [
+		Icons({
+			compiler: 'svelte',
+			autoInstall: true
+		}),
+		sveltekit(),
+		tailwindcss()
+	],
+	server: {
+		proxy: {
+			'/realms': {
+				target: 'http://127.0.0.1:9090',
+				changeOrigin: true
+			},
+			'/resources': {
+				target: 'http://127.0.0.1:9090',
+				changeOrigin: true
+			},
+			'/api': {
+				target: 'http://127.0.0.1:8080',
+				changeOrigin: true
+			}
+		}
+	}
 });
